[330. Patching Array][Compile Error]committed by LeetCode Extension

+++ 330-Patching-Array/solution.java
@@ -1,6 +1,6 @@
 public class Solution {
     public int minPatches(int[] nums, int n) {
-        int known=1, count=0, i=0;
+        long known=1, count=0, i=0;
         while(known<=n)
         {
             if(i<nums.length && known>=nums[i])
